Update Descartes.sol
allow "fully decentralized challenging": if only one party (= the claimer) is specified, anyone can challenge. At first glance, this would not interfere with the "confirm()" function because the check in line 500 always fails if called, although the code should more explicitly disable this function in this case. Not sure how if anything else would break though.